Move environment specific data files
