+++ tests/TestEnvironments/Php81/Rule/CompositeTest.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Yiisoft\Validator\Tests\TestEnvironments\Php81\Rule;
 
 use PHPUnit\Framework\TestCase;
-use Yiisoft\Validator\Tests\Support\Data\CompositeWithCallbackAttribute;
+use Yiisoft\Validator\Tests\TestEnvironments\Support\Data\CompositeWithCallbackAttribute;
 use Yiisoft\Validator\Validator;
 
 final class CompositeTest extends TestCase
