+++ tests/TestEnvironments/Php81/Rule/NestedTest.php
@@ -8,9 +8,9 @@ use PHPUnit\Framework\TestCase;
 use ReflectionProperty;
 use Yiisoft\Validator\Rule\Nested;
 use Yiisoft\Validator\Rule\Number;
-use Yiisoft\Validator\Tests\Support\Data\NestedClassAttribute;
-use Yiisoft\Validator\Tests\Support\Data\NestedWithCallbackAttribute;
 use Yiisoft\Validator\Tests\Support\Data\ObjectWithDifferentPropertyVisibility;
+use Yiisoft\Validator\Tests\TestEnvironments\Support\Data\NestedClassAttribute;
+use Yiisoft\Validator\Tests\TestEnvironments\Support\Data\NestedWithCallbackAttribute;
 use Yiisoft\Validator\Validator;
 
 final class NestedTest extends TestCase
