+++ tests/TestEnvironments/Php81/Rule/StopOnErrorTest.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Yiisoft\Validator\Tests\TestEnvironments\Php81\Rule;
 
 use PHPUnit\Framework\TestCase;
-use Yiisoft\Validator\Tests\Support\Data\StopOnErrorDto;
+use Yiisoft\Validator\Tests\TestEnvironments\Support\Data\StopOnErrorDto;
 use Yiisoft\Validator\Validator;
 
 final class StopOnErrorTest extends TestCase
